backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
networkx==3.2.1
aiofiles==23.2.1
python-magic-bin==0.4.14
chardet==5.2.0
google-generativeai>=0.8.0
python-dotenv==1.0.0



